Move RequestLogMiddleware from/to service_kit.{logging,api}
