+++ service_kit/logging/__init__.py
@@ -1,4 +1,4 @@
 from .log_level import LogLevel
 from .security_risk import SecurityRisk
-from ._logger import configure_logger, intercept_preconfigured_loggers, logger, RequestLogMiddleware
+from ._logger import configure_logger, intercept_preconfigured_loggers, logger
 from .error_logging import log_basic_error, log_exception_group, log_structured_error
