+++ service_kit/logging/_logger.py
@@ -10,17 +10,34 @@ from pathlib import Path
 from typing import Any
 
 import loguru
-from fastapi import Request, Response
 from loguru import logger as _logger
 from pygments import highlight
 from pygments.formatters import TerminalTrueColorFormatter
 from pygments.lexers import JsonLexer
-from starlette.datastructures import Headers
-from starlette.middleware.base import BaseHTTPMiddleware
 
 from service_kit import BaseModel
 
 from . import LogLevel
+
+
+class InterceptHandler(logging.Handler):
+    # Reroutes messages for Python's logging library to Loguru.
+    # This is useful for catching the log messages from third-party libraries.
+    def emit(self, record: logging.LogRecord) -> None:
+        # Get corresponding Loguru level if it exists.
+        level: str | int
+        try:
+            level = logger.level(record.levelname).name
+        except ValueError:
+            level = record.levelno
+
+        # Find caller from where originated the logged message.
+        frame, depth = inspect.currentframe(), 0
+        while frame and (depth == 0 or frame.f_code.co_filename == logging.__file__):
+            frame = frame.f_back
+            depth += 1
+
+        logger.opt(depth=depth, exception=record.exc_info).log(level, record.getMessage())
 
 
 class Serializer:
@@ -72,87 +89,6 @@ serializer = Serializer(io_stream.isatty())
 # The logger has to be patched before it's imported by anything else, otherwise
 # the importing module won't have access to the patched logger.
 logger = _logger.patch(serializer)
-
-
-class InterceptHandler(logging.Handler):
-    # Reroutes messages for Python's logging library to Loguru.
-    # This is useful for catching the log messages from third-party libraries.
-    def emit(self, record: logging.LogRecord) -> None:
-        # Get corresponding Loguru level if it exists.
-        level: str | int
-        try:
-            level = logger.level(record.levelname).name
-        except ValueError:
-            level = record.levelno
-
-        # Find caller from where originated the logged message.
-        frame, depth = inspect.currentframe(), 0
-        while frame and (depth == 0 or frame.f_code.co_filename == logging.__file__):
-            frame = frame.f_back
-            depth += 1
-
-        logger.opt(depth=depth, exception=record.exc_info).log(level, record.getMessage())
-
-
-# TODO: Move to api??
-class RequestLogMiddleware(BaseHTTPMiddleware):
-    debug: bool = False
-
-    async def dispatch(self, request, call_next):
-        with logger.contextualize(request_id=request.state.id):
-            await RequestLogMiddleware.log_request(request)
-            response = await call_next(request)
-            await RequestLogMiddleware.log_response(response)
-
-        return response
-
-    @classmethod
-    async def log_request(cls, request: Request) -> None:
-        try:
-            request_body = await request.json()
-        except json.JSONDecodeError:
-            request_body = str(await request.body())
-
-        common_request_fields = {
-            "method": request.method,
-            "path": request.url.path,
-            "query_parameters": dict(request.query_params),
-            "source": request.client._asdict() if request.client is not None else None,
-            "url": str(request.url),
-        }
-
-        sanitized_headers = RequestLogMiddleware.sanitize_headers(request.headers)
-
-        if cls.debug:
-            logger.debug(
-                "Request received",
-                **common_request_fields,
-                headers=sanitized_headers,
-                body=request_body,
-            )
-        logger.info("Request received", **common_request_fields)
-
-    @staticmethod
-    def sanitize_headers(headers: Headers) -> dict[str, str]:
-        sanitized_headers = {}
-
-        for k, v in headers.items():
-            if k.lower() == "authorization":
-                sanitized_headers[k] = "********"
-            else:
-                sanitized_headers[k] = v
-
-        return sanitized_headers
-
-    @classmethod
-    async def log_response(cls, response: Response):
-        if cls.debug:
-            logger.debug(
-                "Sending response",
-                headers=dict(response.headers),
-                status_code=response.status_code,
-            )
-        logger.info("Sending reponse", status_code=response.status_code)
 
 
 def configure_logger(
